+++ marshmallow/serializer.py
@@ -8,11 +8,11 @@ from marshmallow import base, exceptions, fields, utils
 from marshmallow.compat import with_metaclass, iteritems, text_type, OrderedDict
 
 
-def _get_declared_fields(bases, attrs):
+def _get_declared_fields(bases, attrs, field_class):
     '''Return the declared fields of a class as an OrderedDict.'''
     declared = [(field_name, attrs.pop(field_name))
                 for field_name, val in list(iteritems(attrs))
-                if utils.is_instance_or_subclass(val, base.FieldABC)]
+                if utils.is_instance_or_subclass(val, field_class)]
     # If subclassing another Serializer, inherit its fields
     # Loop in reverse to maintain the correct field order
     for base_class in bases[::-1]:
@@ -28,7 +28,7 @@ class SerializerMeta(type):
     '''
 
     def __new__(cls, name, bases, attrs):
-        attrs['_base_fields'] = _get_declared_fields(bases, attrs)
+        attrs['_base_fields'] = _get_declared_fields(bases, attrs, base.FieldABC)
         return super(SerializerMeta, cls).__new__(cls, name, bases, attrs)